requirements.txt
# Core dependencies
pydantic==2.11.7
httpx==0.28.1
openai==1.97.1
python-dotenv==1.1.1
ichatbio-sdk==0.2.2
instructor==1.10.0
typing_extensions
requests  
cloudscraper    
PyYAML

# LangChain for ReAct agent
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.23.8
pytest-cov==4.1.0
pytest-httpx==0.35.0
pytest-xdist==3.8.0
pytest-watch==4.2.0